backend/requirements.txt
django==5.0.1
channels==4.0.0
djangorestframework==3.14.0
python-jose[cryptography]==3.3.0
uvicorn[standard]==0.25.0